refactor: move portfolio-validation to CSV service directory
Move orphaned portfolio-validation module from utils/ to csv/ service
directory where it belongs.

Changes:
- Move src/lib/utils/portfolio-validation.ts → src/lib/services/csv/portfolio-validation.ts
- Move src/lib/utils/__tests__/portfolio-validation.test.ts → src/lib/services/csv/__tests__/portfolio-validation.test.ts
- Update import in csv-import-dialog.tsx

Rationale:
- portfolio-validation.ts only used by CSV import (single consumer)
- Located in general-purpose utils/ directory (incorrect location)
- CSV-specific validation logic should be in CSV service directory

Impact: Improved organization, CSV-specific code now in CSV directory

Test Results: All 11 tests passed in new location ✅
